requirements.txt
customtkinter==5.2.2
Pillow==11.3.0
mysql-connector-python==9.4.0 